Move data objects to pinot-controller
