+++ pinot-tools/src/main/java/org/apache/pinot/tools/realtime/provisioning/MemoryEstimator.java
@@ -34,11 +34,11 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.io.FileUtils;
-import org.apache.pinot.common.data.schema.DateTimeFieldSpecMetadata;
-import org.apache.pinot.common.data.schema.FieldMetadata;
-import org.apache.pinot.common.data.schema.SchemaWithMetaData;
-import org.apache.pinot.common.data.schema.TimeFieldSpecMetadata;
-import org.apache.pinot.common.data.schema.TimeGranularitySpecMetadata;
+import org.apache.pinot.controller.recommender.io.metadata.DateTimeFieldSpecMetadata;
+import org.apache.pinot.controller.recommender.io.metadata.FieldMetadata;
+import org.apache.pinot.controller.recommender.io.metadata.SchemaWithMetaData;
+import org.apache.pinot.controller.recommender.io.metadata.TimeFieldSpecMetadata;
+import org.apache.pinot.controller.recommender.io.metadata.TimeGranularitySpecMetadata;
 import org.apache.pinot.common.metadata.segment.RealtimeSegmentZKMetadata;
 import org.apache.pinot.common.segment.ReadMode;
 import org.apache.pinot.core.data.readers.PinotSegmentRecordReader;
